backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
pydantic 
aiofiles
boto3
python-dotenv
authlib
requests
werkzeug
email_validator
pydantic-settings